backend/requirements.txt
anyio==3.6.2
bcrypt==4.0.1
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
fastapi==0.92.0
h11==0.14.0
idna==3.4
joblib==1.2.0
nltk==3.8.1
numpy==1.21.6
pandas==1.3.5
passlib==1.7.4
pydantic==1.10.5
PyJWT==2.6.0
pymongo==4.3.3
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.7.1
regex==2022.10.31
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
tqdm==4.64.1
typing==3.7.4.3
typing_extensions==4.5.0
uvicorn==0.20.0
